Bitacora Asegurado, Permiso, Rol, Permiso, RolPermiso

+++ src/app/business/asegurado/asegurado.component.ts
@@ -12,6 +12,10 @@ import { CalendarModule } from 'primeng/calendar';
 import { MessageService } from 'primeng/api';
 import { ButtonModule } from 'primeng/button';
 import { RippleModule } from 'primeng/ripple';
+import { BitacoraService } from '../../core/services/bitacora.service'; // Importa el servicio de bitácora
+import { Bitacora } from '../../core/models/bitacora';
+import { AuthService } from '../../core/services/auth.service';
+
 
 @Component({
   selector: 'app-asegurado',
@@ -49,7 +53,9 @@ export default class AseguradoComponent {
   constructor(
     private aseguradoService: AseguradoService,
     private usuarioService: UsuarioService,
-    private messageService: MessageService
+    private messageService: MessageService,
+    private bitacoraService: BitacoraService,
+    private authService: AuthService
   ) {}
 
   ngOnInit(): void {
@@ -64,6 +70,32 @@ export default class AseguradoComponent {
     );
   }
 
+  registrarBitacora(accion: string, detalle: string): void {
+    this.bitacoraService.getUserIP().subscribe({
+      next: (response) => {
+        const now = new Date();
+        const fecha = `${now.getFullYear()}-${(now.getMonth() + 1).toString().padStart(2, '0')}-${now.getDate().toString().padStart(2, '0')}`;
+        const hora = `${now.getHours().toString().padStart(2, '0')}:${now.getMinutes().toString().padStart(2, '0')}:${now.getSeconds().toString().padStart(2, '0')}`;
+  
+        const bitacoraEntry: Bitacora = {
+          correo: this.authService.getAuthenticatedUserEmail() || '',
+          fecha: fecha,
+          hora: hora,
+          ip: response.ip,
+          accion: accion,
+          detalle: detalle
+        };
+  
+        this.bitacoraService.createBitacora(bitacoraEntry).subscribe({
+          next: () => console.log('Registro de bitácora exitoso'),
+          error: (err) => console.error('Error al registrar en bitácora', err)
+        });
+      },
+      error: (err) => console.error('Error al obtener IP', err)
+    });
+  }
+  
+
   getUsuariosDisponibles() {
     this.usuarioService.getUsuariosSinAseguradoByRol(2).subscribe(
       (data) => (this.usuariosDisponibles = data),
@@ -77,6 +109,8 @@ export default class AseguradoComponent {
         this.getAsegurados()
         this.asegurados.push(data);
         this.messageService.add({ severity: 'success', summary: 'Añadido', detail: 'Asegurado añadido correctamente' });
+        // Registro en bitácora
+        this.registrarBitacora('Añadir asegurado', `Asegurado creado: ${data.usuario?.correo}`);
         this.resetNuevoAsegurado();
         this.getUsuariosDisponibles();
       },
@@ -95,10 +129,18 @@ export default class AseguradoComponent {
   }
 
   deleteAsegurado(id: number) {
+    // Guardamos el asegurado antes de eliminarlo
+    const aseguradoEliminado = this.asegurados.find((asegurado) => asegurado.id === id);
+    if (!aseguradoEliminado) {
+      this.messageService.add({ severity: 'error', summary: 'Error', detail: 'Asegurado no encontrado' });
+      return;
+    }
     this.aseguradoService.deleteAsegurado(id).subscribe(
       () => {
         this.asegurados = this.asegurados.filter((asegurado) => asegurado.id !== id);
         this.messageService.add({ severity: 'success', summary: 'Eliminado', detail: 'Asegurado eliminado correctamente' });
+        // Registro en bitácora
+        this.registrarBitacora('Eliminar asegurado', `Asegurado eliminado: ${aseguradoEliminado?.usuario?.correo}`);
         this.getUsuariosDisponibles();
       },
       (error) => this.messageService.add({ severity: 'error', summary: 'Error', detail: 'No se pudo eliminar el asegurado' })
@@ -114,6 +156,8 @@ export default class AseguradoComponent {
     this.aseguradoService.updateAsegurado(asegurado).subscribe(() => {
       delete this.editedAsegurados[asegurado.id!];
       this.messageService.add({ severity: 'success', summary: 'Guardado', detail: 'Tipo de sangre actualizado' });
+      // Registro en bitácora
+      this.registrarBitacora('Modificar asegurado', `Asegurado modificado: ${asegurado.usuario?.correo}`);
     });
   }
